refactor location of tests
