+++ .gitignore
@@ -3,3 +3,6 @@ testing-infra/external/movement-aptos-core/
 
 testing-infra/single-validator/work/data/
 testing-infra/.aptos/
+
+# Temporary e2e test entry point (generated by run-tests.sh)
+trusted-verifier/tests/integration_test_e2e.rs
