+++ README.md
@@ -16,12 +16,12 @@ nix develop
 ### Testing
 
 ```
+# Unit tests (no Docker required)
 cd move-intent-framework && aptos move test --dev --named-addresses aptos_intent=0x123 && cd ..
-
-# the following test is also required for the trusted-verifier integration test
-./trusted-verifier/tests/integration/run-cross-chain-verifier.sh 1
-
 cd trusted-verifier && cargo test --locked && cd ..
+
+# E2E integration tests (requires Docker)
+./testing-infra/e2e-tests/complete-system/run-tests.sh
 ```
 
 ## License
