+++ testing-infra/e2e-tests/complete-system/connectivity_test.rs
