+++ testing-infra/e2e-tests/complete-system/event_polling_test.rs
