+++ testing-infra/e2e-tests/complete-system/mod.rs
@@ -2,6 +2,9 @@
 //!
 //! This module contains integration tests that require external services
 //! (such as running Aptos chains).
+//!
+//! These tests are located in testing-infra/e2e-tests/complete-system/
+//! and require Docker chains to be running.
 
 pub mod connectivity_test;
 pub mod deployment_test;
