+++ testing-infra/e2e-tests/complete-system/run-tests.sh
@@ -1,0 +1,102 @@
+#!/bin/bash
+
+# E2E Integration Test Runner
+# 
+# This script runs the Rust integration tests that require Docker chains.
+# It sets up chains, deploys contracts, submits intents, then runs the tests.
+
+set -e
+
+# Get the project root
+SCRIPT_DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
+PROJECT_ROOT="$( cd "$SCRIPT_DIR/../../.." && pwd )"
+cd "$PROJECT_ROOT"
+
+echo "🧪 E2E Integration Tests Runner"
+echo "================================"
+echo ""
+
+echo "🚀 Step 0: Setting up chains, deploying contracts, and submitting intents..."
+echo "========================================================================"
+./testing-infra/e2e-tests/move-intent-framework/submit-cross-chain-intent.sh 1
+
+if [ $? -ne 0 ]; then
+    echo "❌ Failed to setup and deploy contracts"
+    exit 1
+fi
+
+echo ""
+echo "✅ Setup complete! Extracting module addresses..."
+
+# Extract deployed addresses from aptos profiles and update verifier.toml
+CHAIN1_ADDRESS=$(aptos config show-profiles | jq -r '.["Result"]["intent-account-chain1"].account')
+CHAIN2_ADDRESS=$(aptos config show-profiles | jq -r '.["Result"]["intent-account-chain2"].account')
+
+if [ -z "$CHAIN1_ADDRESS" ] || [ -z "$CHAIN2_ADDRESS" ]; then
+    echo "❌ ERROR: Could not extract deployed module addresses"
+    exit 1
+fi
+
+echo "   Chain 1 deployer: $CHAIN1_ADDRESS"
+echo "   Chain 2 deployer: $CHAIN2_ADDRESS"
+
+# Update verifier.toml
+VERIFIER_CONFIG="trusted-verifier/config/verifier.toml"
+if [ ! -f "$VERIFIER_CONFIG" ]; then
+    echo "❌ ERROR: verifier.toml not found at $VERIFIER_CONFIG"
+    exit 1
+fi
+
+# Update module addresses in verifier.toml
+sed -i "/\[hub_chain\]/,/\[connected_chain\]/ s|intent_module_address = .*|intent_module_address = \"0x$CHAIN1_ADDRESS\"|" "$VERIFIER_CONFIG"
+sed -i "/\[connected_chain\]/,/\[verifier\]/ s|intent_module_address = .*|intent_module_address = \"0x$CHAIN2_ADDRESS\"|" "$VERIFIER_CONFIG"
+sed -i "/\[connected_chain\]/,/\[verifier\]/ s|escrow_module_address = .*|escrow_module_address = \"0x$CHAIN2_ADDRESS\"|" "$VERIFIER_CONFIG"
+
+# Get Alice and Bob addresses and update known_accounts
+ALICE_CHAIN1_ADDRESS=$(aptos config show-profiles | jq -r '.["Result"]["alice-chain1"].account')
+BOB_CHAIN1_ADDRESS=$(aptos config show-profiles | jq -r '.["Result"]["bob-chain1"].account')
+ALICE_CHAIN2_ADDRESS=$(aptos config show-profiles | jq -r '.["Result"]["alice-chain2"].account')
+
+if [ -n "$ALICE_CHAIN1_ADDRESS" ] && [ -n "$BOB_CHAIN1_ADDRESS" ]; then
+    sed -i "/\[hub_chain\]/,/\[connected_chain\]/ s|known_accounts = .*|known_accounts = [\"$ALICE_CHAIN1_ADDRESS\", \"$BOB_CHAIN1_ADDRESS\"]|" "$VERIFIER_CONFIG"
+fi
+
+if [ -n "$ALICE_CHAIN2_ADDRESS" ]; then
+    sed -i "/\[connected_chain\]/,/\[verifier\]/ s|known_accounts = .*|known_accounts = [\"$ALICE_CHAIN2_ADDRESS\"]|" "$VERIFIER_CONFIG"
+fi
+
+echo "✅ Updated verifier.toml with deployed addresses"
+echo ""
+
+echo "📋 Running E2E integration tests..."
+echo ""
+
+# Create a temporary integration test entry point
+INTEGRATION_TEST_FILE="trusted-verifier/tests/integration_test_e2e.rs"
+cat > "$INTEGRATION_TEST_FILE" << 'EOF'
+//! E2E Integration tests entry point
+//!
+//! This file is temporarily generated by run-tests.sh to load e2e tests.
+
+#[path = "../../testing-infra/e2e-tests/complete-system/mod.rs"]
+mod integration;
+EOF
+
+# Run the tests from trusted-verifier directory
+cd trusted-verifier
+cargo test --test integration_test_e2e --locked
+
+# Clean up temporary test file
+cd ..
+rm -f "$INTEGRATION_TEST_FILE"
+
+echo ""
+echo "✅ E2E integration tests completed!"
+echo ""
+echo "🚀 Step 2: Running verifier service to test end-to-end flow..."
+echo "============================================================"
+./testing-infra/e2e-tests/complete-system/run-cross-chain-verifier.sh 0
+
+echo ""
+echo "✅ All E2E tests completed!"
+
