+++ testing-infra/e2e-tests/move-intent-framework/setup-and-deploy.sh
