--- trusted-verifier/tests/integration_test.rs
@@ -1,7 +1,0 @@
-//! Integration tests entry point
-//!
-//! This file is discovered by Cargo and loads all integration tests
-//! from the `tests/integration/` subdirectory.
-
-mod integration;
-
